fix(build): unify dynamic segment names under admin users
