+++ app/api/admin/users/[id]/subscription/route.ts
@@ -1,0 +1,19 @@
+import { NextRequest, NextResponse } from 'next/server';
+
+export const dynamic = 'force-dynamic';
+
+const ADMIN_SECRET = process.env.ADMIN_SECRET || 'changeme';
+
+export async function PUT(request: NextRequest, { params }: { params: { id: string } }) {
+  const adminKey = request.headers.get('x-admin-secret');
+  if (adminKey !== ADMIN_SECRET) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
+  const { id } = params;
+  const { status = 'free', expiresAt } = await request.json();
+
+  console.log(`Admin updated user ${id} subscription to ${status}`, expiresAt);
+
+  return NextResponse.json({ success: true, userId: id, status, expiresAt });
+}
